Send dashboard job messages in chunks
Reduce the rate of the messages going out to the dashboard
send messages in packets of N every M seconds where N
and M are configurable for all messages. The DashboardReport
spawns a thread to run the handleCreated and handleStatusChanged
methods.

Also add a load test to use in coordination with the dashboard people
to tune the rate of the messages.

Fixes #4053