Improve TaskArchiver cleanup
This fixes the lockout bug due to the parentage check
in the files, the query should be a little faster now.

The parentage check is still enforced in the Subscription.deleteEverything
function, the file deletion is extended to ensure that all records
are deleted.

Fixes #4341
Fixes #4405 - (Likely, not worth to keep that one open).